Task #383: stabilize quote/invoice test helpers
