+++ backend/app/features/quotes/tests/support/__init__.py
@@ -1,0 +1,1 @@
+"""Shared test support for quote/invoice API integration tests."""
