+++ backend/app/features/quotes/tests/support/helpers.py
@@ -1,0 +1,282 @@
+from __future__ import annotations
+
+from datetime import date
+from uuid import UUID, uuid4
+
+from httpx import AsyncClient
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
+
+from app.features.auth.models import User
+from app.features.auth.service import CSRF_COOKIE_NAME
+from app.features.quotes.models import Document, QuoteStatus
+from app.worker.job_registry import extraction_job, pdf_job
+from app.worker.runtime import (
+    DEFAULT_MAX_TRIES,
+    DEFAULT_RETRY_BASE_SECONDS,
+    DEFAULT_RETRY_JITTER_SECONDS,
+    WorkerRuntimeSettings,
+)
+
+from .mocks import (
+    _MockExtractionIntegration,
+    _MockPdfIntegration,
+    _MockStorageService,
+)
+
+
+def _send_email_headers(csrf_token: str, *, idempotency_key: str | None = None) -> dict[str, str]:
+    return {
+        "X-CSRF-Token": csrf_token,
+        "Idempotency-Key": idempotency_key or uuid4().hex,
+    }
+
+
+def _assert_async_email_job_response(response, *, document_id: str) -> dict[str, object]:
+    assert response.status_code == 202
+    payload = response.json()
+    assert payload["job_type"] == "email"
+    assert payload["status"] == "pending"
+    assert payload["document_id"] == document_id
+    return payload
+
+
+async def _register_and_login(client: AsyncClient, credentials: dict[str, str]) -> str:
+    register_response = await client.post("/api/auth/register", json=credentials)
+    assert register_response.status_code == 201
+
+    login_response = await client.post("/api/auth/login", json=credentials)
+    assert login_response.status_code == 200
+    csrf_token = client.cookies.get(CSRF_COOKIE_NAME)
+    assert csrf_token is not None
+    return csrf_token
+
+
+async def _create_customer(
+    client: AsyncClient,
+    csrf_token: str,
+    *,
+    name: str = "Quote Test Customer",
+    email: str | None = None,
+    phone: str | None = None,
+    address: str | None = None,
+) -> str:
+    response = await client.post(
+        "/api/customers",
+        json={
+            "name": name,
+            "email": email,
+            "phone": phone,
+            "address": address,
+        },
+        headers={"X-CSRF-Token": csrf_token},
+    )
+    assert response.status_code == 201
+    return response.json()["id"]
+
+
+async def _create_quote(client: AsyncClient, csrf_token: str, customer_id: str) -> dict[str, str]:
+    response = await client.post(
+        "/api/quotes",
+        json={
+            "customer_id": customer_id,
+            "transcript": "quote transcript",
+            "line_items": [{"description": "line item", "details": None, "price": 55}],
+            "total_amount": 55,
+            "notes": "Original note",
+            "source_type": "text",
+        },
+        headers={"X-CSRF-Token": csrf_token},
+    )
+    assert response.status_code == 201
+    return response.json()
+
+
+async def _create_direct_invoice(
+    client: AsyncClient,
+    csrf_token: str,
+    customer_id: str,
+    *,
+    title: str | None,
+    transcript: str,
+    total_amount: int,
+) -> dict[str, object]:
+    response = await client.post(
+        "/api/invoices",
+        json={
+            "customer_id": customer_id,
+            "title": title,
+            "transcript": transcript,
+            "line_items": [{"description": "line item", "details": None, "price": total_amount}],
+            "total_amount": total_amount,
+            "notes": "Original note",
+            "source_type": "text",
+        },
+        headers={"X-CSRF-Token": csrf_token},
+    )
+    assert response.status_code == 201
+    return response.json()
+
+
+async def _create_approved_invoice(
+    client: AsyncClient,
+    csrf_token: str,
+    db_session: AsyncSession,
+) -> dict[str, object]:
+    customer_id = await _create_customer(client, csrf_token)
+    quote = await _create_quote(client, csrf_token, customer_id)
+    await _set_quote_status(db_session, quote["id"], QuoteStatus.APPROVED)
+
+    response = await client.post(
+        f"/api/quotes/{quote['id']}/convert-to-invoice",
+        headers={"X-CSRF-Token": csrf_token},
+    )
+    assert response.status_code == 201
+    return response.json()
+
+
+async def _set_profile_for_email_delivery(client: AsyncClient, csrf_token: str) -> None:
+    response = await client.patch(
+        "/api/profile",
+        json={
+            "business_name": "Summit Exterior Care",
+            "first_name": "Jane",
+            "last_name": "Doe",
+            "trade_type": "Landscaper",
+            "timezone": "America/New_York",
+        },
+        headers={"X-CSRF-Token": csrf_token},
+    )
+    assert response.status_code == 200
+
+
+async def _set_user_phone_number(
+    db_session: AsyncSession,
+    *,
+    email: str,
+    phone_number: str,
+) -> None:
+    user = await _get_user_by_email(db_session, email)
+    user.phone_number = phone_number
+    await db_session.commit()
+
+
+async def _set_user_email_and_phone_number(
+    db_session: AsyncSession,
+    *,
+    email: str,
+    updated_email: str,
+    phone_number: str | None,
+) -> None:
+    user = await _get_user_by_email(db_session, email)
+    user.email = updated_email
+    user.phone_number = phone_number
+    await db_session.commit()
+
+
+async def _get_user_by_email(db_session: AsyncSession, email: str) -> User:
+    user = await db_session.scalar(select(User).where(User.email == email))
+    assert user is not None
+    return user
+
+
+async def _set_quote_status(
+    db_session: AsyncSession,
+    quote_id: str,
+    status: QuoteStatus,
+) -> None:
+    quote = await db_session.scalar(select(Document).where(Document.id == UUID(quote_id)))
+    assert quote is not None
+    quote.status = status
+    await db_session.commit()
+
+
+async def _set_invoice_status(
+    db_session: AsyncSession,
+    invoice_id: object,
+    status: QuoteStatus,
+) -> None:
+    assert isinstance(invoice_id, str)
+    invoice = await db_session.scalar(select(Document).where(Document.id == UUID(invoice_id)))
+    assert invoice is not None
+    invoice.status = status
+    await db_session.commit()
+
+
+async def _run_pdf_job(db_session: AsyncSession, *, job_id: object) -> None:
+    assert isinstance(job_id, str)
+    session_maker = async_sessionmaker(
+        bind=db_session.bind,
+        class_=AsyncSession,
+        expire_on_commit=False,
+    )
+    runtime = WorkerRuntimeSettings(
+        session_maker=session_maker,
+        max_tries=DEFAULT_MAX_TRIES,
+        retry_base_seconds=DEFAULT_RETRY_BASE_SECONDS,
+        retry_jitter_seconds=DEFAULT_RETRY_JITTER_SECONDS,
+    )
+    await pdf_job(
+        {
+            "job_try": 1,
+            "worker_runtime": runtime,
+            "pdf_integration": _MockPdfIntegration(),
+            "storage_service": _MockStorageService(),
+        },
+        job_id,
+    )
+
+
+async def _run_extraction_job(
+    db_session: AsyncSession,
+    *,
+    job_id: object,
+    source_type: str,
+    capture_detail: str,
+    customer_id: str | None = None,
+    append_to_quote: bool = False,
+    transcript: str = "mulch the front beds",
+    job_try: int = 1,
+    extraction_integration: object | None = None,
+    correlation_id: str | None = None,
+) -> None:
+    assert isinstance(job_id, str)
+    session_maker = async_sessionmaker(
+        bind=db_session.bind,
+        class_=AsyncSession,
+        expire_on_commit=False,
+    )
+    runtime = WorkerRuntimeSettings(
+        session_maker=session_maker,
+        max_tries=DEFAULT_MAX_TRIES,
+        retry_base_seconds=DEFAULT_RETRY_BASE_SECONDS,
+        retry_jitter_seconds=DEFAULT_RETRY_JITTER_SECONDS,
+    )
+    resolved_extraction_integration = extraction_integration or _MockExtractionIntegration()
+    await extraction_job(
+        {
+            "job_try": job_try,
+            "worker_runtime": runtime,
+            "extraction_integration": resolved_extraction_integration,
+        },
+        job_id,
+        correlation_id=correlation_id,
+        transcript=transcript,
+        source_type=source_type,
+        capture_detail=capture_detail,
+        customer_id=customer_id,
+        append_to_quote=append_to_quote,
+    )
+
+
+def _format_human_date(value: object) -> str:
+    assert isinstance(value, str)
+    return date.fromisoformat(value).strftime("%b %d, %Y").replace(" 0", " ")
+
+
+def _credentials() -> dict[str, str]:
+    suffix = uuid4().hex[:12]
+    return {
+        "email": f"user-{suffix}@example.com",
+        "password": "StrongPass123!",
+    }
